requirements.txt
pandas>=1.4.1
numpy>=1.22.3
scikit-learn>=1.0.2
lightgbm>=3.3.2
xgboost>=1.6.1
imbalanced-learn>=0.9.0
hydra-core>=1.2.0
mlflow>=2.0.0
pyyaml>=6.0
omegaconf>=2.2.1
joblib>=1.1.0
shap>=0.41.0
matplotlib>=3.5.1
seaborn>=0.11.2
tqdm
pyarrow
click>=8.1.0
rich>=12.0.0
questionary>=1.10.0
pytest
streamlit>=1.10.0
plotly>=5.7.0
matplotlib-venn
ipython
catboost>=1.0.6
openpyxl>=3.0.9
PyYAML>=6.0
optuna>=3.0.0
colorlog>=6.6.0

